added activites table
